backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
workos==5.29.0
